requirements.txt
Flask>=2.3.2
Flask_Caching>=2.0.2
Requests>=2.31.0
mutagen>=1.46.0
Pillow~=10.1.0